# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HR Management App API Documentation",
        "contact": {
            "email": "alabaganne9@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/collaborators/{user_id}": {
            "get": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Collaborator details",
                "description": "Get collaborator informations",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "Collaborator details response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "$ref": "#/components/schemas/user"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Edit collaborator",
                "description": "Edit an existing collaborator",
                "operationId": "store",
                "requestBody": {
                    "$ref": "#/components/requestBodies/collaboratorRequestBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/invalid-data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Delete collaborator",
                "description": "Delete a collaborator",
                "operationId": "App\\Http\\Controllers\\Collaborators\\CollaboratorController::destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/user_id"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/user_id"
                }
            ]
        },
...
        "/api/collaborators/{user_id}/leaves/{leave_id}": {
            "put": {
                "tags": [
                    "Leaves"
                ],
                "summary": "Update user leave",
                "description": "Update specified user leave",
                "operationId": "App\\Http\\Controllers\\Collaborators\\LeaveController::update",
                "requestBody": {
                    "$ref": "#/components/requestBodies/leaveRequestBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Leaves"
                ],
                "summary": "Delete a specified leave from storage",
                "description": "Delete a specified leave from storage",
                "operationId": "App\\Http\\Controllers\\Collaborators\\LeaveController::destroy",
                "requestBody": {
                    "$ref": "#/components/requestBodies/leaveRequestBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/user_id"
                },
                {
                    "$ref": "#/components/parameters/leave_id"
                }
            ]
        },
...
        "/api/collaborators/{user_id}/skills/{skill_id}": {
            "put": {
                "tags": [
                    "Skills"
                ],
                "summary": "Edit user skill",
                "description": "Update a skill for a user",
                "operationId": "App\\Http\\Controllers\\Collaborators\\SkillController::update",
                "requestBody": {
                    "$ref": "#/components/requestBodies/skillRequestBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/invalid-data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Skills"
                ],
                "summary": "Revoke a skill from a user",
                "description": "Revoke a skill to a user and delete it if no user still have it",
                "operationId": "App\\Http\\Controllers\\Collaborators\\SkillController::destroy",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/user_id"
                },
                {
                    "$ref": "#/components/parameters/skill_id"
                }
            ]
        },
...
        "/api/collaborators": {
            "post": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Get list of collaborators",
                "description": "Get paginated list of collaborators",
                "operationId": "index",
                "requestBody": {
                    "$ref": "#/components/requestBodies/searchRequestBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/paginatedCollaborators"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/invalid-data"
                    }
                }
            }
        },
        "/api/collaborators/create": {
            "post": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Add a new collaborator",
                "description": "Create a new collaborator",
                "operationId": "store",
                "requestBody": {
                    "$ref": "#/components/requestBodies/collaboratorRequestBody"
                },
                "responses": {
                    "201": {
                        "description": "Collaborator created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "collaborator_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/invalid-data"
                    }
                }
            }
        },
        "/api/collaborators/gender": {
            "get": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Number of collaborators per gender",
                "description": "Number of collaborators for each gender",
                "operationId": "App\\Http\\Controllers\\Collaborators\\CollaboratorController::collaboratorsNumberByGender",
                "responses": {
                    "200": {
                        "description": "Number of collaborators per gender",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "male": {
                                            "type": "integer",
                                            "example": 7
                                        },
                                        "female": {
                                            "type": "integer",
                                            "example": 6
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
        "/api/collaborators/department": {
            "get": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Number of collaborators per department",
                "description": "Number of collaborators for each department",
                "operationId": "App\\Http\\Controllers\\Collaborators\\CollaboratorController::collaboratorsNumberByDepartment",
                "responses": {
                    "200": {
                        "description": "Number of collaborators per department",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Web": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "Mobile": {
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "AI": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "Data Science": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "Design": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
        "/api/collaborators/archive": {
            "get": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Collaborators from the archive",
                "description": "Collaborators from the archive",
                "operationId": "App\\Http\\Controllers\\Collaborators\\CollaboratorController::archive",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/paginatedCollaborators"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
        "/api/collaborators/{user_id}/restore": {
            "get": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Restore a collaborator",
                "description": "Restore a collaborator",
                "operationId": "App\\Http\\Controllers\\Collaborators\\CollaboratorController::restore",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/user_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User restored."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
        "/api/collaborators/{user_id}/delete-permantly": {
            "delete": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Delete a collaborator permantly",
                "description": "Delete a collaborator permantly from the database",
                "operationId": "App\\Http\\Controllers\\Collaborators\\CollaboratorController::deletePermantly",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/user_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User permantly deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "user": {
                "title": "User",
                "description": "User object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Ala Baganne"
                    },
                    "username": {
                        "type": "string",
                        "example": "alabaganne"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "alabaganne9@gmail.com"
                    },
                    "password": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "50101959"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "dateTime",
                        "example": "1999-12-16"
                    },
                    "address": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female"
                        ]
                    },
                    "civil_status": {
                        "enum": [
                            "single",
                            "married"
                        ]
                    },
                    "id_card_number": {
                        "type": "string"
                    },
                    "nationality": {
                        "type": "string",
                        "example": "Tunisian"
                    },
                    "university": {
                        "type": "string",
                        "example": "Higher institute of applied science and technology of sousse"
                    },
                    "history": {
                        "type": "string"
                    },
                    "experience_level": {
                        "type": "integer",
                        "maximum": 5,
                        "example": 4
                    },
                    "source": {
                        "type": "string"
                    },
                    "position": {
                        "type": "string",
                        "example": "Full stack web developer"
                    },
                    "grade": {
                        "type": "string"
                    },
                    "hiring_date": {
                        "type": "string",
                        "format": "dateTime"
                    },
                    "contract_end_date": {
                        "type": "string",
                        "format": "dateTime"
                    },
                    "type_of_contract": {
                        "type": "string",
                        "enum": [
                            "option 1",
                            "option 2",
                            "option 3"
                        ]
                    },
                    "allowed_leave_days": {
                        "type": "integer",
                        "default": 30
                    },
                    "department_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "unauthorized": {
                "description": "Forbidden",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "This action is unauthorized."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "unauthenticated": {
                "description": "Authentication is required",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
...
        "requestBodies": {
...
            "collaboratorRequestBody": {
                "description": "Collaborator request body",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/user"
                        }
                    }
                }
            },
...
        }
    },
...
    "security": [
        []
    ]
}